ci(gdn-tp): file-size split (tp_shard/gdn.rs) + fmt + clippy/typos green
- Split the GDN HeadParallel slicers out of tp_shard.rs into tp_shard/gdn.rs
  (704 -> 343 LoC; both under the 500 cap) — drops tp_shard.rs from the
  file-size allow-list; allow-list linear_attn_arms.rs (548) + trait_decode_batched.rs (503).
- clippy: repeat().take() -> std::iter::repeat_n() in tp_shard/tests.rs.
- typos: alog_src_off -> a_log_src_off.
- cargo fmt.
