+++ .github/workflows/file-size-cap.yml
@@ -212,6 +212,15 @@ jobs:
             # max-batch-size + FP8-KV flags. Marginally over; back under 500 in
             # the follow-on dense-FP8 PR. Clean split tracked as follow-up.
             "crates/spark-server/src/cli/serve_args.rs"
+            # 548 LoC — Qwen3.5 linear-attn (SSM) loader arm; grew with the
+            # HeadParallel config-localize + per-rank segmented weight loads
+            # (FP8-dense + NVFP4 builders). Two linear load sequences read
+            # top-to-bottom; clean split tracked as follow-up.
+            "crates/spark-model/src/weight_loader/qwen35/load_layers/linear_attn_arms.rs"
+            # 503 LoC — batched-recurrent GDN decode; marginally over (added the
+            # per-seq SSM out_proj all-reduce for HeadParallel). Back under 500
+            # once the batched/graph paths consolidate. Follow-up.
+            "crates/spark-model/src/layers/qwen3_ssm/trait_decode_batched.rs"
           )
 
           violations=0
